Implement health system and refactor architecture for improved maintainability
- Completed health system integration, including player health management, invincibility, and game over transitions.
- Removed the unused SystemManager and related code, streamlining the architecture.
- Deleted the entire internal physics package and fixed various lint violations to enhance code quality.
- Refactored game.go by splitting it into focused modules, optimizing file sizes and improving clarity.
- Achieved zero linter issues while modernizing the codebase to align with current best practices.
